pyproject.toml
[project]
name = "memco"
version = "1.0.0"
authors = [
    { name = "Arubik Dev" },
]
description = "A comprehensive memory management system with vector search capabilities"
readme = "README.md"
requires-python = ">=3.7"
keywords = ["memory", "vector search", "embeddings", "ai"]
classifiers = [
        "Programming Language :: Python :: 3",
        "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
        "cryptography",
        "python-dotenv",
        "requests",
        "numpy",
        "rich",
        "fastapi",
        "uvicorn",
        "ttkbootstrap"
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"

[project.scripts]
memco = "memco:cli_main"
